day01-jwt-basics/auth-service/requirements.txt
# === Сервер ===
fastapi==0.115.0
uvicorn==0.30.6

# === База данных ===
sqlalchemy==2.0.35
psycopg2-binary==2.9.10

# === Настройки ===
pydantic-settings==2.6.0
python-dotenv==1.0.1

# === Аутентификация ===
argon2-cffi==23.1.0
PyJWT==2.9.0

# === Alembic ===
alembic==1.18.3
